Moving utils from test -> src
