+++ modules/passkey/src/utils/safe.ts
