+++ modules/passkey/test/4337/WebAuthnSingletonSigner.spec.ts
@@ -8,7 +8,7 @@ import {
   buildRpcUserOperationFromSafeUserOperation,
 } from '@safe-global/safe-4337/src/utils/userOp'
 import { buildSignatureBytes } from '@safe-global/safe-4337/src/utils/execution'
-import { WebAuthnCredentials, decodePublicKey, encodeWebAuthnSignature } from '../utils/webauthn'
+import { WebAuthnCredentials, decodePublicKey, encodeWebAuthnSignature } from '../../src/utils/webauthn'
 
 describe('WebAuthn Singleton Signers [@4337]', () => {
   before(function () {
