+++ modules/passkey/test/GasBenchmarking.spec.ts
@@ -1,7 +1,7 @@
 import { expect } from 'chai'
 import { deployments, ethers } from 'hardhat'
 
-import { WebAuthnCredentials, decodePublicKey, encodeWebAuthnSignature } from './utils/webauthn'
+import { WebAuthnCredentials, decodePublicKey, encodeWebAuthnSignature } from '../src/utils/webauthn'
 import { IP256Verifier } from '../typechain-types'
 
 describe('Gas Benchmarking [@bench]', function () {
