+++ modules/passkey/test/userstories/PasskeyCredentialCreation.spec.ts
@@ -1,8 +1,8 @@
 import { expect } from 'chai'
 import { deployments, ethers } from 'hardhat'
-import { WebAuthnCredentials, decodePublicKey, encodeWebAuthnSignature } from '../utils/webauthn'
+import { WebAuthnCredentials, decodePublicKey, encodeWebAuthnSignature } from '../../src/utils/webauthn'
 import { buildSignatureBytes } from '@safe-global/safe-4337/src/utils/execution'
-import { buildSafeTransaction, buildSafeTransactionData, SafeDomain } from '../utils/safe'
+import { buildSafeTransaction, buildSafeTransactionData, SafeDomain } from '../../src/utils/safe'
 
 /**
  * User story: Passkey Credential Creation for Safe Ownership
